requirements.txt
python-telegram-bot[webhooks]==20.5
requests==2.31.0